+++ hooks/android/sound/utils.js
@@ -1,7 +1,0 @@
-module.exports = {
-  getAppName: function (context) {
-    let ConfigParser = context.requireCordovaModule("cordova-lib").configparser;
-    let config = new ConfigParser("config.xml");
-    return config.name();
-  }
-};