+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
+
+### 04-08-2023
+- Feat: Trigger NotificationClick for the Deprecated Block  (https://outsystemsrd.atlassian.net/browse/RMET-2736)
+- Feat: Trigger NotificationClick for the Supported Block  (https://outsystemsrd.atlassian.net/browse/RMET-2738)
+
 - Feat: update sound hook to unzip sound files, for both iOS and Android (https://outsystemsrd.atlassian.net/browse/RMET-2464).
 - Feat: update firebase core version (https://outsystemsrd.atlassian.net/browse/RMET-2451).